lambda-architecture-demo/requirements.txt
certifi
faker
flask
flask_script
kafka-python
requests
azure-eventhub